app/requirements.txt
requests
pandas
psycopg2-binary
scikit-learn